careeros/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "careeros"
version = "0.1.0"
description = "CareerOS — Application Intelligence Pack Generator"
requires-python = ">=3.11"

dependencies = [
    # Web framework
    "fastapi==0.111.0",
    "uvicorn[standard]==0.29.0",

    # Database
    "sqlalchemy[asyncio]==2.0.30",
    "asyncpg==0.29.0",
    "alembic==1.13.1",

    # Validation
    "pydantic==2.7.1",
    "pydantic-settings==2.2.1",

    # Queue
    "celery==5.4.0",
    "redis==5.0.4",

    # Storage
    "boto3==1.34.100",

    # Document processing
    "python-docx==1.1.2",
    "python-multipart==0.0.9",

    # LLM
    "anthropic==0.26.1",

    # HTTP client (retrieval service + Stripe)
    "httpx==0.27.0",
    "stripe==9.9.0",  # Optional: can use httpx client instead

    # Utilities
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "structlog==24.1.0",
    "sentry-sdk[fastapi]==2.3.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.2.0",
    "pytest-asyncio==0.23.6",
    "pytest-cov==5.0.0",
    "httpx==0.27.0",
    "ruff==0.4.4",
    "mypy==1.10.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
